requirements.txt
Flask==1.1.1
request==2019.4.13
slackclient==2.5.0
